Internals

QuantumCitations.CitationBibliography — Type

Plugin for enabling APS style citations in Documenter.jl.

bib = CitationBibliography(filename; sorting=:none)

instantiates a plugin that must be passed as an (undocumented!) positional argument to Documenter.makedocs.

Internally, bib can then be accessed as doc.plugins[CitationBibliography] where doc is the Documenter.Documents.Document object.

Internal Attributes

filename: name filename with which bib was instantiated

bib: the BibTeX data, as obtained by BibParser.parse_file

citations: an ordered dict mapping citation keys to their numerical citation key

page_citations: a dict mapping page file names (md files inside docs/src) to a set of citation keys that are cited on that page

The QuantumCitations.CitationBibliography plugin hooks into the Documenter.Builder.DocumentPipeline[1] between ExpandTemplates (which expands @docs blocks) and CrossReferences. The plugin adds the following three steps:

CollectCitations

ExpandBibliography

ExpandCitations

QuantumCitations.CollectCitations — Type

Pipeline step to collect citations from all pages.

It walks all pages in the order they appear in the navigation bar, looking for @cite links. It fills the citations and page_citations attributes of the internal CitationBibliography object.

Thus, the order in which CollectCitations encounters citations determines the numerical key that will appear in the rendered documentation (see ExpandBibliography and ExpandCitations).

QuantumCitations.ExpandBibliography — Type

Pipeline step to expand all @bibliography blocks.

Runs after CollectCitations but before ExpandCitations.

Each bibliography is rendered into HTML as a definition list. The "term" for each list item (the numerical citation key) is rendered via format_bibliography_key and the "description" (the actual bibliographic reference) is rendered via format_bibliography_entry.

QuantumCitations.ExpandCitations — Type

Pipeline step to expand all @cite citations.

This runs after ExpandBibliography, as it relies on the link targets in the expanded @bibliography blocks.

All citations are formatted using format_citation.

Customization

Even though QuantumCitations targets the APS/REVTeX numeric citation style, it is technically possible to completely customize the rendering of citations and references by overwriting the format_bibliography_key, format_bibliography_entry, and format_citation methods detailed below.

QuantumCitations.format_bibliography_key — Function

Format the key for an entry in a @bibliography block.

format_bibliography_key(entry, doc)

produces a string for the rendered key in the bibliography for the given Bibliography.Entry

It determines a numerical citation key by looking up entry.id in doc.plugins[CitationBibliography].citations. This numerical key is returned in square brackets.

If overridden, this method should generally match format_citation.

QuantumCitations.format_bibliography_entry — Function

Format an entry in a @bibliography block.

format_bibliography_entry(entry)

produces an HTML string from a Bibliography.Entry that is formatted like in REVTeX and APS journals. That is, the full list of authors with initials for the first names, the italicized tile, and the journal reference (linking to the DOI, if available), ending with the publication year in parenthesis.

QuantumCitations.format_citation — Function

Format a @cite citation.

Returns a string that replaces a [key](@cite) citation in markdown text.

This returns the same numerical key in square brackets that the default format_bibliography_key produces in the rendered bibliography.

Debugging

Set the environment variable JULIA_DEBUG=Documenter,QuantumCitations before generating the documentation.

1See the documentation of Documenter.Utilities.Selectors for an explanation of Documenter's pipeline concept.